pulling things specific to wx ButtonEditor into the implementation file and out of default
